+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] -
 
+### Added
+
+- `address` field to `/addresses/{address}/utxos` and `/addresses/{address}/utxos/{asset}`
+
+### Changed
+
+- increased token registry timeout to 5s -> 10s
+
+### Fixed
+
+- `/addresses/{address}/utxos/{asset}` asset validation for `/lovelace`
+
 ## [1.1.1] - 2022-12-07
 
 ### Fixed